refactor(critical): extract MatchWidget test fixtures to satisfy 500-line limit (ARC-678)
MatchWidget.test.tsx had grown to 520 lines after ARC-673 added the
localStorage-toggle suite. Move the makeProps factory into a sibling
fixtures module so the test file drops to 462 lines. vi.mock factories
stay inline because Vitest hoists them per-file.
