+++ apps/web/src/widgets/CriticalGame/ui/MatchWidget.test.tsx
@@ -168,65 +168,7 @@ vi.mock('./styles/layout', () => ({
 
 import { MatchWidget, type MatchWidgetProps } from './MatchWidget';
 import type { CriticalCard, CriticalPlayerState } from '../types';
-
-function makeProps(override: Partial<MatchWidgetProps> = {}): MatchWidgetProps {
-  const currentPlayer: CriticalPlayerState = {
-    playerId: 'p1',
-    hand: ['strike', 'strike', 'evade'] as CriticalCard[],
-    alive: true,
-  };
-  const baseSnapshot = {
-    deck: [] as CriticalCard[],
-    discardPile: [] as CriticalCard[],
-    playerOrder: ['p1'],
-    currentTurnIndex: 0,
-    pendingDraws: 1,
-    pendingDefuse: null,
-    pendingFavor: null,
-    pendingAlter: null,
-    pendingAction: null,
-    players: [currentPlayer],
-    logs: [],
-    allowActionCardCombos: true,
-  };
-  return {
-    room: { id: 'room' } as never,
-    snapshot: baseSnapshot,
-    currentUserId: 'p1',
-    currentPlayer,
-    cardVariant: 'cyberpunk',
-    isGameOver: false,
-    isMyTurn: true,
-    canAct: true,
-    canPlayNope: false,
-    actionBusy: null,
-    aliveOpponents: [],
-    handLayout: 'grid',
-    setHandLayout: vi.fn(),
-    resolveDisplayName: (_id?: string, fb?: string) => fb,
-    t: ((k: string) => k) as never,
-    actions: {
-      drawCard: vi.fn(),
-      playNope: vi.fn(),
-      playSeeTheFuture: vi.fn(),
-      playFavor: vi.fn(),
-      giveFavorCard: vi.fn(),
-      playDefuse: vi.fn(),
-      playActionCard: vi.fn(),
-    } as never,
-    idleTimerTriggered: false,
-    autoplayState: { setAllEnabled: vi.fn() },
-    handleUnstash: vi.fn(),
-    handlePlayActionCard: vi.fn(),
-    handleOpenFavorModal: vi.fn(),
-    handleOpenEventCombo: vi.fn(),
-    handleOpenFiverCombo: vi.fn(),
-    formatLogMessage: (m?: string | null) => m ?? '',
-    isFullscreen: false,
-    toggleFullscreen: vi.fn(),
-    ...override,
-  } as MatchWidgetProps;
-}
+import { makeProps } from './MatchWidget.test-fixtures';
 
 describe('MatchWidget (ARC-635)', () => {
   beforeEach(() => {
